IGNITE-11945: Metric API changes.
